+++ src/API/Types.hs
@@ -851,6 +851,10 @@ data QualityReportAPI = QualityReportAPI
     , qraTruncatedNameCollisions :: QualityCheckAPI
     , qraMissingPedigree :: QualityCheckAPI
     , qraUnconsumedProducts :: QualityCheckAPI
+    , qraLandTransformationBalance :: QualityCheckAPI
+    , qraOxygenDemandOrder :: QualityCheckAPI
+    , qraInvalidCas :: QualityCheckAPI
+    , qraAllocationOutOfRange :: QualityCheckAPI
     }
     deriving (Generic)
     deriving (ToJSON, FromJSON, ToSchema) via (Stripped QualityReportAPI)